Move checkrun logic to ChecksClientWrapper
